refactor: densify e2e dismiss/move mirror harnesses (Refs #4075)
Extract shared WidgetTester hosts under test/support/ and migrate dismiss_*
plus pick-move-destination suites (Slice C / AC5–6).
