+++ packages/colonizethis_app_e2e_support/test/e2e_pick_move_destination_and_confirm_test.dart
@@ -31,8 +31,6 @@ library;
 // Same pattern as `e2e_radio_list_tiles_in_alert_dialogs_test.dart`.
 // ignore_for_file: deprecated_member_use
 
-import 'package:colonizethis_app_fixtures/config/ct_e2e.dart';
-import 'package:colonizethis_app_l10n/l10n/app_localizations_contract.dart';
 import 'package:colonizethis_app_l10n/l10n/app_localizations_lookup.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
@@ -40,237 +38,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app_e2e_support/e2e_test_shared.dart';
 
-/// Single source of truth for the warp row text under English locale.
-/// Matches `l10n.moveFleet_warpLinkToRegion(regionDisplayLabel('newWorld'))`
-/// — see `app/lib/l10n/app_localizations_en_part2.dart` and
-/// `app/lib/features/game/widgets/units/shared/region_labels.dart`.
-const String _warpText = 'links to New World';
-
-/// Sea-zone destination row label distinct from the warp text so the
-/// fixture exercises the multi-row branch (warp present vs sea-only) without
-/// ambiguity.
-const String _seaText = 'sea zone 1';
-
-/// Test-only host that renders a `MoveFleetDialog`-shaped [AlertDialog].
-///
-/// The dialog uses a [Scrollable] keyed by
-/// [kCtE2EMoveFleetDialogScrollRootKey] when [withScrollKey] is true (the
-/// production [MoveFleetDialog] body always carries this key). When false,
-/// the helper falls back to the descendant [Scrollable] under
-/// [AlertDialog].
-class _MoveDialogHost extends StatefulWidget {
-  const _MoveDialogHost({
-    super.key,
-    required this.l10n,
-    required this.includeWarp,
-    this.includeSea = true,
-    this.withScrollKey = true,
-    this.warpAheadFillers = 0,
-    this.scrollPhysics,
-    this.includeScrollable = true,
-    this.wrapWarpInIgnorePointer = false,
-  });
-
-  final AppLocalizations l10n;
-  final bool includeWarp;
-  final bool includeSea;
-  final bool withScrollKey;
-
-  /// When > 0, inserts that many tall filler rows BEFORE the warp row so the
-  /// row is initially off-screen and the helper must drag the scrollable to
-  /// reveal it.
-  final int warpAheadFillers;
-
-  /// Custom scroll physics for the dialog's [SingleChildScrollView]. When set
-  /// to [NeverScrollableScrollPhysics] the [tester.drag] gesture cannot
-  /// produce a scroll offset, though [tester.scrollUntilVisible] (which
-  /// composes [Scrollable.ensureVisible] programmatically) still bypasses
-  /// physics.
-  final ScrollPhysics? scrollPhysics;
-
-  /// When false, the dialog content is a bare [Column] with no [Scrollable];
-  /// the helper's `find.descendant(of: AlertDialog, matching: Scrollable)`
-  /// branch resolves empty and the warp-scroll loop is skipped entirely.
-  final bool includeScrollable;
-
-  /// When true, wraps the warp [RadioListTile] (and its `Text(warpText)`
-  /// title) in an [IgnorePointer], so the Text widget remains findable via
-  /// `find.textContaining(...)` but `warp.hitTestable()` always resolves
-  /// empty. Forces the helper's deterministic `'Warp row not hit-testable'`
-  /// failure path even when [tester.scrollUntilVisible] would otherwise
-  /// reveal the row.
-  final bool wrapWarpInIgnorePointer;
-
-  @override
-  State<_MoveDialogHost> createState() => _MoveDialogHostState();
-}
-
-class _MoveDialogHostState extends State<_MoveDialogHost> {
-  int? selectedRowIndex;
-  int taps = 0;
-  bool dialogOpen = true;
-
-  @override
-  Widget build(BuildContext context) {
-    if (!dialogOpen) {
-      return const SizedBox.shrink();
-    }
-    final rows = <Widget>[];
-    final indices = <int, String>{};
-    var nextIndex = 0;
-    for (var i = 0; i < widget.warpAheadFillers; i++) {
-      rows.add(
-        SizedBox(
-          height: 200,
-          child: Center(child: Text('filler-$i')),
-        ),
-      );
-    }
-    if (widget.includeWarp) {
-      final warpIndex = nextIndex++;
-      indices[warpIndex] = 'warp';
-      final Widget warpTile = RadioListTile<int>(
-        key: const ValueKey('warp-tile'),
-        title: const Text(_warpText),
-        value: warpIndex,
-        groupValue: selectedRowIndex,
-        onChanged: (next) {
-          taps++;
-          setState(() => selectedRowIndex = next);
-        },
-      );
-      rows.add(
-        widget.wrapWarpInIgnorePointer
-            ? IgnorePointer(child: warpTile)
-            : warpTile,
-      );
-    }
-    if (widget.includeSea) {
-      final seaIndex = nextIndex++;
-      indices[seaIndex] = 'sea';
-      rows.add(
-        RadioListTile<int>(
-          key: const ValueKey('sea-tile'),
-          title: const Text(_seaText),
-          value: seaIndex,
-          groupValue: selectedRowIndex,
-          onChanged: (next) {
-            taps++;
-            setState(() => selectedRowIndex = next);
-          },
-        ),
-      );
-    }
-    final Widget content;
-    if (widget.includeScrollable) {
-      content = SingleChildScrollView(
-        key: widget.withScrollKey ? kCtE2EMoveFleetDialogScrollRootKey : null,
-        physics: widget.scrollPhysics,
-        child: Column(mainAxisSize: MainAxisSize.min, children: rows),
-      );
-    } else {
-      content = Column(mainAxisSize: MainAxisSize.min, children: rows);
-    }
-    return AlertDialog(
-      content: SizedBox(
-        width: 320,
-        height: 240,
-        child: content,
-      ),
-      actions: <Widget>[
-        TextButton(
-          onPressed: () {
-            setState(() => dialogOpen = false);
-          },
-          child: Text(widget.l10n.common_confirm),
-        ),
-      ],
-    );
-  }
-
-  String? get selectedKind {
-    final i = selectedRowIndex;
-    if (i == null) {
-      return null;
-    }
-    var nextIndex = 0;
-    if (widget.includeWarp) {
-      if (i == nextIndex) {
-        return 'warp';
-      }
-      nextIndex++;
-    }
-    if (widget.includeSea) {
-      if (i == nextIndex) {
-        return 'sea';
-      }
-    }
-    return null;
-  }
-}
-
-Future<void> _pumpDialogScaffold(
-  WidgetTester tester, {
-  required Widget child,
-}) async {
-  await tester.pumpWidget(
-    MaterialApp(
-      home: Scaffold(
-        body: Builder(
-          builder: (context) => Center(
-            child: SizedBox.expand(child: child),
-          ),
-        ),
-      ),
-    ),
-  );
-}
-
-/// Counter for unique [_MoveDialogHost] keys so each `_pumpMoveDialog` call
-/// in the same `testWidgets` body mounts a fresh State (no `dialogOpen=false`
-/// leakage between invocations of [e2ePickMoveDestinationAndConfirm]).
-int _hostKeyCounter = 0;
-
-Future<_MoveDialogHostState> _pumpMoveDialog(
-  WidgetTester tester, {
-  required AppLocalizations l10n,
-  required bool includeWarp,
-  bool includeSea = true,
-  bool withScrollKey = true,
-  int warpAheadFillers = 0,
-  ScrollPhysics? scrollPhysics,
-  bool includeScrollable = true,
-  bool wrapWarpInIgnorePointer = false,
-}) async {
-  _hostKeyCounter++;
-  final hostKey = ValueKey<int>(_hostKeyCounter);
-  late _MoveDialogHostState hostState;
-  await tester.pumpWidget(
-    MaterialApp(
-      home: Scaffold(
-        body: Builder(
-          builder: (context) {
-            return Center(
-              child: _MoveDialogHost(
-                key: hostKey,
-                l10n: l10n,
-                includeWarp: includeWarp,
-                includeSea: includeSea,
-                withScrollKey: withScrollKey,
-                warpAheadFillers: warpAheadFillers,
-                scrollPhysics: scrollPhysics,
-                includeScrollable: includeScrollable,
-                wrapWarpInIgnorePointer: wrapWarpInIgnorePointer,
-              ),
-            );
-          },
-        ),
-      ),
-    ),
-  );
-  hostState = tester.state<_MoveDialogHostState>(find.byType(_MoveDialogHost));
-  return hostState;
-}
+import 'support/move_dialog_widget_tester_harness.dart';
 
 void main() {
   suppressLogsForTests();
@@ -280,7 +48,7 @@ void main() {
       'taps warp RadioListTile when present and confirms (default args)',
       (WidgetTester tester) async {
         final l10n = lookupAppLocalizations(const Locale('en'));
-        final host = await _pumpMoveDialog(
+        final host = await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: true,
@@ -323,7 +91,7 @@ void main() {
       'taps warp tile via RadioListTile ancestor (not the inner Text)',
       (WidgetTester tester) async {
         final l10n = lookupAppLocalizations(const Locale('en'));
-        final host = await _pumpMoveDialog(
+        final host = await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: true,
@@ -351,13 +119,13 @@ void main() {
       'drag-probes the scrollable when warp row is initially off-screen',
       (WidgetTester tester) async {
         final l10n = lookupAppLocalizations(const Locale('en'));
-        final host = await _pumpMoveDialog(
+        final host = await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: true,
           warpAheadFillers: 5,
         );
-        final warpFinder = find.text(_warpText);
+        final warpFinder = find.text(kMoveDialogWarpText);
         expect(
           warpFinder.evaluate().isNotEmpty,
           isTrue,
@@ -392,7 +160,7 @@ void main() {
         // its own because [tester.scrollUntilVisible] composes
         // [Scrollable.ensureVisible] programmatically and bypasses scroll
         // physics.
-        await _pumpMoveDialog(
+        await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: true,
@@ -434,7 +202,7 @@ void main() {
       'is absent',
       (WidgetTester tester) async {
         final l10n = lookupAppLocalizations(const Locale('en'));
-        final host = await _pumpMoveDialog(
+        final host = await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: true,
@@ -462,7 +230,7 @@ void main() {
       'taps sea RadioListTile when warp row is absent',
       (WidgetTester tester) async {
         final l10n = lookupAppLocalizations(const Locale('en'));
-        final host = await _pumpMoveDialog(
+        final host = await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: false,
@@ -496,7 +264,7 @@ void main() {
         // ordering puts the on-screen warp tile first but the helper's
         // sea-radio branch picks `seaRadio.first` directly — proving no
         // warp-scroll work happened).
-        final host = await _pumpMoveDialog(
+        final host = await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: true,
@@ -562,7 +330,7 @@ void main() {
         // sea-radio branch and confirm without depending on
         // `kCtE2EMoveFleetDialogScrollRootKey` or a descendant
         // [Scrollable].
-        final host = await _pumpMoveDialog(
+        final host = await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: false,
@@ -607,7 +375,7 @@ void main() {
       'fails with deterministic message when no AlertDialog mounts in time',
       (WidgetTester tester) async {
         final l10n = lookupAppLocalizations(const Locale('en'));
-        await _pumpDialogScaffold(tester, child: const SizedBox());
+        await pumpMoveDialogScaffold(tester, child: const SizedBox());
 
         Object? caught;
         try {
@@ -647,7 +415,7 @@ void main() {
       'fails on the very first ensureBudget when moveDialogBudget is zero',
       (WidgetTester tester) async {
         final l10n = lookupAppLocalizations(const Locale('en'));
-        await _pumpMoveDialog(
+        await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: true,
@@ -687,7 +455,7 @@ void main() {
       (WidgetTester tester) async {
         final l10n = lookupAppLocalizations(const Locale('en'));
 
-        final hostA = await _pumpMoveDialog(
+        final hostA = await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: true,
@@ -696,7 +464,7 @@ void main() {
         expect(hostA.selectedKind, 'warp');
         expect(hostA.dialogOpen, isFalse);
 
-        final hostB = await _pumpMoveDialog(
+        final hostB = await pumpMoveDialog(
           tester,
           l10n: l10n,
           includeWarp: true,
